pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bedrock-llm-comparison"
version = "1.0.0"
description = "Enterprise multi-LLM comparison tool for AWS Bedrock, OpenAI, and Gemini"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Organization", email = "contact@yourorg.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]

dependencies = [
    "boto3>=1.34.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "openai>=1.0.0",
    "google-genai>=0.5.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pylint>=3.0.0",
    "black>=23.0.0",
]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.pylint.messages_control]
max-line-length = 100
disable = ["C0111", "C0103"]
